obs: add test for remote upload, make it pass
The test exposed the fact that ssh and scp were invoked
nonuniformly; the fix was to always invoke them the same
way.  Use shellcheck-approved quoting on the changed lines.

Also, in the spirit of paranoid error checking, strategically
add trailing slashes on scp commands to fail if detination
is accidentally a file.